설문 수정

+++ src/main/java/com/freeneo/survey/mapper/SurveyMapper.java
@@ -6,8 +6,8 @@ import com.freeneo.survey.domain.Survey;
 
 public interface SurveyMapper {
 	public List<Survey> list();
-	public Survey select(int id);
+	public Survey select(Long id);
 	public int insert(Survey survey);
 	public int update(Survey survey);
-	public int delete(int id);
+	public int delete(Long id);
 }
